% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_a_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Convierte Temperatura de Celsius a Fahrenheit}
\usage{
celsius_to_fahrenheit(temp_centigrados)
}
\arguments{
\item{temp_centigrados}{Una temperatura en grado Celsius.}
}
\value{
La temperatura ingresada pero en grados Fahrenheit.
}
\description{
La función \code{celsius_to_fahrenheit} toma una temperatura en grados Celsius y la convierte en grados Fahrenheit.
}
